Rename mark renderers
