+++ sg2d-wgpu/src/marks/mod.rs
@@ -6,5 +6,5 @@ pub mod rect;
 pub mod rule;
 pub mod symbol;
 pub mod text;
-pub mod texture_instanced_mark;
-pub mod texture_mark;
+pub mod instanced_mark;
+pub mod basic_mark;
